Feat(newTalks): reposicionamento de arquivos e uso de jsons
