+++ src/bot/commands/compet/newTalks/interactionReply.json
@@ -1,0 +1,4 @@
+{
+    "content": "Titulo do Forms Alterado. Link do forms: https://docs.google.com/forms/d/e/1FAIpQLSeh2F82mKd8t6RAXuXdKU7kQiGxButW8xNK8FZSF4wCO-ZuHQ/viewform",
+    "ephemeral": true
+}
