+++ src/bot/commands/compet/newTalks/new-talks-embed-error.json
